requirements.txt
# Core dependencies
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7

# AI and Translation
openai==1.3.7
google-generativeai==0.8.4
edge-tts==6.1.9

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Security
PyJWT==2.8.0